[FE] REFACTOR: divide components and container #846
