+++ frontend/src/components/LeftNav/LeftMainNav/LeftMainNav.tsx
@@ -1,103 +1,22 @@
-import React, { useEffect } from "react";
 import styled from "styled-components";
-import { useLocation, useNavigate } from "react-router-dom";
-import {
-  useRecoilState,
-  useRecoilValue,
-  useSetRecoilState,
-  useResetRecoilState,
-} from "recoil";
-import {
-  currentFloorNumberState,
-  currentFloorCabinetState,
-  currentSectionNameState,
-  currentLocationNameState,
-  userState,
-  isCurrentSectionRenderState,
-} from "@/recoil/atoms";
-import { currentLocationFloorState } from "@/recoil/selectors";
-import { axiosCabinetByLocationFloor } from "@/api/axios/axios.custom";
-import { CabinetInfoByLocationFloorDto } from "@/types/dto/cabinet.dto";
-import { removeCookie } from "@/api/react_cookie/cookies";
-import useIsMount from "@/hooks/useIsMount";
-import { UserDto } from "@/types/dto/user.dto";
 
-const LeftMainNav = () => {
-  const floors = useRecoilValue<Array<number>>(currentLocationFloorState);
-  const [currentFloor, setCurrentFloor] = useRecoilState<number>(
-    currentFloorNumberState
-  );
-  const currentLocation = useRecoilValue<string>(currentLocationNameState);
-  const myInfo = useRecoilValue<UserDto>(userState);
-  const resetCurrentFloor = useResetRecoilState(currentFloorNumberState);
-  const resetCurrentSection = useResetRecoilState(currentSectionNameState);
-  const resetLocation = useResetRecoilState(currentLocationNameState);
-  const setCurrentFloorData = useSetRecoilState<
-    CabinetInfoByLocationFloorDto[]
-  >(currentFloorCabinetState);
-  const setCurrentSection = useSetRecoilState<string>(currentSectionNameState);
-  const navigator = useNavigate();
-  const { pathname } = useLocation();
-  const isMount = useIsMount();
-  const [isCurrentSectionRender, setIsCurrentSectionRender] = useRecoilState(
-    isCurrentSectionRenderState
-  );
+interface ILeftMainNav {
+  pathname: string;
+  onClickHomeButton: React.MouseEventHandler;
+  floors: number[];
+  currentFloor: number;
+  onClickFloorButton: Function;
+  onClickLogoutButton: React.MouseEventHandler;
+}
 
-  useEffect(() => {
-    if (currentFloor === undefined) return;
-    axiosCabinetByLocationFloor(currentLocation, currentFloor)
-      .then((response) => {
-        setCurrentFloorData(response.data);
-        if (isMount || isCurrentSectionRender) {
-          const recoilPersist = localStorage.getItem("recoil-persist");
-          let recoilPersistObj;
-          if (recoilPersist) recoilPersistObj = JSON.parse(recoilPersist);
-          setCurrentSection(
-            Object.keys(recoilPersistObj).includes("CurrentSection")
-              ? recoilPersistObj.CurrentSection
-              : response.data[0].section
-          );
-          setIsCurrentSectionRender(false);
-        } else {
-          setCurrentSection(response.data[0].section);
-        }
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  }, [currentLocation, currentFloor, myInfo.cabinet_id]);
-
-  const onClickFloorButton = (floor: number) => {
-    setCurrentFloor(floor);
-    if (pathname == "/home") {
-      if (floor === currentFloor) {
-        axiosCabinetByLocationFloor(currentLocation, currentFloor).then(
-          (response) => {
-            setCurrentFloorData(response.data);
-            setCurrentSection(response.data[0].section);
-          }
-        );
-      }
-      navigator("/main");
-    }
-  };
-
-  const onClickHomeButton = () => {
-    navigator("/home");
-  };
-
-  const onClickLogoutButton = (): void => {
-    if (import.meta.env.VITE_IS_LOCAL === "true") {
-      removeCookie("access_token");
-    } else {
-      removeCookie("access_token", { path: "/", domain: "cabi.42seoul.io" });
-    }
-    resetLocation();
-    resetCurrentFloor();
-    resetCurrentSection();
-    navigator("/login");
-  };
-
+const LeftMainNav = ({
+  pathname,
+  floors,
+  currentFloor,
+  onClickHomeButton,
+  onClickFloorButton,
+  onClickLogoutButton,
+}: ILeftMainNav) => {
   return (
     <LeftNavStyled>
       <TopSectionStyled>
@@ -108,19 +27,20 @@ const LeftMainNav = () => {
           >
             Home
           </TopBtnStyled>
-          {floors.map((floor, index) => (
-            <TopBtnStyled
-              className={
-                pathname != "/home" && floor === currentFloor
-                  ? "leftNavButtonActive"
-                  : ""
-              }
-              onClick={() => onClickFloorButton(floor)}
-              key={index}
-            >
-              {floor + "층"}
-            </TopBtnStyled>
-          ))}
+          {floors &&
+            floors.map((floor, index) => (
+              <TopBtnStyled
+                className={
+                  pathname != "/home" && floor === currentFloor
+                    ? "leftNavButtonActive"
+                    : ""
+                }
+                onClick={() => onClickFloorButton(floor)}
+                key={index}
+              >
+                {floor + "층"}
+              </TopBtnStyled>
+            ))}
         </TopBtnsStyled>
       </TopSectionStyled>
       <BottomSectionStyled>
