+++ frontend/src/components/LeftNav/LeftNav.tsx
@@ -1,13 +1,13 @@
 import styled from "styled-components";
-import LeftMainNav from "@/components/LeftNav/LeftMainNav/LeftMainNav";
-import LeftSectionNav from "@/components/LeftNav/LeftSectionNav/LeftSectionNav";
+import LeftMainNavContainer from "./LeftMainNav/LeftMainNav.container";
+import LeftSectionNavContainer from "./LeftSectionNav/LeftSectionNav.container";
 
 const LeftNav: React.FC<{ isVisible: boolean }> = (props) => {
   return (
     <>
       <LeftNavWrapStyled id="leftNavWrap">
-        <LeftMainNav></LeftMainNav>
-        <LeftSectionNav isVisible={props.isVisible}></LeftSectionNav>
+        <LeftMainNavContainer></LeftMainNavContainer>
+        <LeftSectionNavContainer isVisible={props.isVisible} />
       </LeftNavWrapStyled>
     </>
   );
